Move MissingApplicationID to top-level discord namespace
