+++ discord/app_commands/errors.py
@@ -27,7 +27,7 @@ from __future__ import annotations
 from typing import Any, TYPE_CHECKING, List, Optional, Sequence, Union
 
 from ..enums import AppCommandOptionType, AppCommandType, Locale
-from ..errors import DiscordException, HTTPException, _flatten_error_dict
+from ..errors import DiscordException, HTTPException, _flatten_error_dict, MissingApplicationID as MissingApplicationID
 from ..utils import _human_join
 
 __all__ = (
@@ -58,11 +58,6 @@ if TYPE_CHECKING:
     from .checks import Cooldown
 
     CommandTypes = Union[Command[Any, ..., Any], Group, ContextMenu]
-
-APP_ID_NOT_FOUND = (
-    'Client does not have an application_id set. Either the function was called before on_ready '
-    'was called or application_id was not passed to the Client constructor.'
-)
 
 
 class AppCommandError(DiscordException):
@@ -420,19 +415,6 @@ class CommandSignatureMismatch(AppCommandError):
             'command tree to fix this issue.'
         )
         super().__init__(msg)
-
-
-class MissingApplicationID(AppCommandError):
-    """An exception raised when the client does not have an application ID set.
-    An application ID is required for syncing application commands.
-
-    This inherits from :exc:`~discord.app_commands.AppCommandError`.
-
-    .. versionadded:: 2.0
-    """
-
-    def __init__(self, message: Optional[str] = None):
-        super().__init__(message or APP_ID_NOT_FOUND)
 
 
 def _get_command_error(
